+++ src/components/app/Writing.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import { type Snippet } from 'svelte';
     import Page from '@components/app/Page.svelte';
+    import Breadcrumbs from '@components/app/Breadcrumbs.svelte';
 
     interface Props {
         children: Snippet;
@@ -12,6 +13,7 @@
 
 <Page {footer}>
     <div class="writing">
+        <Breadcrumbs />
         {@render children()}
     </div>
 </Page>